Array Manipulation Example
Starting with a 1-indexed array of zeros and a list of operations, for each operation add a value to each the array element between two given indices, inclusive. Once all operations have been performed, return the maximum value in the array.

Sample Input:
5 3
1 2 100
2 5 100
3 4 100

Sample Output:

200

Explanation:

After the first update the list is 100 100 0 0 0.
After the second update list is 100 200 100 100 100.
After the third update list is 100 200 200 200 100.

Then return 200.

arrayManipulationSolutionFromOthers --> This file has been copied from existing solutions in the web. 

 arrayManipulation_solutionOne and arrayManipulation_solutionTwo

Both functions works fine for smaller size arrays, if the number of elements in the array is to large then the memory can be overloaded. 

Everyone allowed to improve solutionOne and solutionTwo performance, feel free to comment

Bye.