Repated String
There is a string, s, of lowercase English letters that is repeated infinitely many times. Given an integer, n, find and print the number of letter a's in the first n letters of the infinite string.

Example:
s = "abcac"
n = 10
The substring we consider is abcacabcacabcacabcac......, the first 10 characters of the infinite string. There are 4 occurrences of a in the substring.

Sample Input:
aba
10
Sample Output:
7

Sample Input:
a
1000000000000
Sample Output:
1000000000000